requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.5
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiter==0.7.0
langfuse==2.53.9
numpy==2.1.3
openai==1.53.1
packaging==24.2
portalocker==2.10.1
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
qdrant-client==1.12.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0
